Some more debugging from client side

+++ src/python/client/client.py
@@ -16,10 +16,9 @@ PLATFORM_INFO_SEP = '_'
 PLATFORM_INFO_OS_ARCH_REL_FMTS = list(map(lambda cortege: PLATFORM_INFO_SEP.join(list(cortege)), itertools.permutations(['{platform}', '{architecture}', '{release}'])))
 PLATFORM_INFO_OS_ARCH_FMTS     = list(map(lambda cortege: PLATFORM_INFO_SEP.join(list(cortege)), itertools.permutations(['{platform}', '{architecture}'])))
 PLATFORM_INFO_OS_REL_FMTS      = list(map(lambda cortege: PLATFORM_INFO_SEP.join(list(cortege)), itertools.permutations(['{platform}', '{release}'])))
-PLATFORM_INFO_OTHER_FMTS       = ['{platform}', '']
+PLATFORM_INFO_OTHER_FMTS       = ['{platform}']
 
 START_SCRIPT_PREFIX = 'start'
-STOP_SCRIPT_PREFIX  = 'stop'
 
 SCRIPT_RESULTS_DIR = 'result'
 SCRIPT_EXTENTIONS = ['exe', 'sh', 'py']
@@ -27,14 +26,15 @@ SCRIPT_EXTENTIONS = ['exe', 'sh', 'py']
 ZIPPER_CMD_FMT   = 'tar -czf {{name}}.{archext} {{wildcard}}'.format(archext = ARCHIVE_EXTENTION)
 UNZIPPER_CMD_FMT = 'tar -xzf {wildcard}'
 
+TRAITS_COLSEP = ' '
+
+TASK_STATUS_FAILED = 'failed'
+TASK_STATUS_FINISHED = 'finished'
+
 config = {}
 
 config.update(GRID_CALC_ROLE = 'GRID_CALC_CLIENT')
-config.update(WORKING_DIRECTORY = os.path.join(tempfile.gettempdir(), '_'.join([config['GRID_CALC_ROLE'], 'work'])))
-config.update(HOME_DIRECTORY = os.path.join(os.path.expanduser("~"), config['GRID_CALC_ROLE']))
-config.update(ARCHIVE_DIRECTORY = os.path.join(tempfile.gettempdir(), '_'.join([config['GRID_CALC_ROLE'], 'arch'])))
 config.update(NODE_FRONTEND = None)
-config.update(LOCK_FILE = 'key.lock')
 config.update(SCRIPT_VALID_POSTFIX_FMTS = PLATFORM_INFO_OS_ARCH_REL_FMTS + PLATFORM_INFO_OS_ARCH_FMTS + PLATFORM_INFO_OS_REL_FMTS + PLATFORM_INFO_OTHER_FMTS)
 config.update(CLIENT_TRAITS = [])
 
@@ -46,6 +46,22 @@ def auto_extract_traits():
     config['CLIENT_TRAITS'] += [dict(name = 'os_version', version = release)]
     config['CLIENT_TRAITS'] += [dict(name = 'architecture', version = arch)]
 
+class EUnrecoverable(Exception):
+    def __init__(self, message):
+        super(EUnrecoverable, self).__init__(message)
+
+class ERecoverable(Exception):
+    def __init__(self, message):
+        super(ERecoverable, self).__init__(message)
+
+class ETaskUnrecoverable(Exception):
+    def __init__(self, message):
+        super(ETaskUnrecoverable, self).__init__(message)
+
+class ETaskRecoverable(Exception):
+    def __init__(self, message):
+        super(ETaskRecoverable, self).__init__(message)
+
 class Node:
     key            = None
     busy           = False
@@ -53,8 +69,11 @@ class Node:
     input_archive  = None
     output_archive = None
     req_timeout    = 1
+    task_req_timeout = 1
     node_id        = None
     task_id        = None
+    max_time       = None
+    status         = None
 
     def __init__(self):
         self.key = random_string(64)
@@ -63,31 +82,31 @@ class Node:
         return ', '.join(map(lambda x: str(getattr(self, x)), filter(lambda x: not hasattr(x, '__call__'))))
 
     def run(self):
-        auto_extract_traits()
-        import pdb
-        pdb.set_trace()
         homedir = config['HOME_DIRECTORY']
         workdir = config['WORKING_DIRECTORY']
         front = 'http://{0}'.format(config['NODE_FRONTEND'])
-        keyfile = os.path.join(homedir, 'key.dat')
-        tasklockfile = os.path.join(homedir, 'task.lock')
+        keyfile = config['KEYFILE']
+        tasklockfile = config['TASK_LOCKFILE']
         
         ###>> keyfile ###
-        if (os.path.exists(os.path.join(keyfile))):
+        if (os.path.exists(keyfile)):
             ifh = open(keyfile, 'r')
             ifs = ifh.read()
             id_, oldkey = None, None
             try:
                 id_, oldkey = ifs.split('#')
             except:
-                raise Exception('Syntax error in {0}. Try to remove it and retart.'.format(keyfile))
+                raise ERecoverable('Syntax error in {0}. Removing it and restarting.'.format(keyfile))
+                os.unlink(keyfile)
             resp = requests.put(front + '/nodes/{0}'.format(id_), data = {'key': self.key, 'key_old': oldkey})
-            assert_response(resp)
+            assert_response(resp, EUnrecoverable)
             self.node_id = id_
         else:
+            auto_extract_traits()
             resp = jsr('post', front + '/nodes', data = {'key': self.key, 'traits': config['CLIENT_TRAITS']})
-            assert_response(resp)
+            assert_response(resp, EUnrecoverable)
             self.node_id = resp.json()['agent_id']
+            self.max_time = resp.json()['max_time']
 
         ofs = open(keyfile, 'w')
         ofs.write('#'.join(map(str, [self.node_id, self.key])))
@@ -102,7 +121,7 @@ class Node:
             resp = requests.get(front + '/tasks/newtask', data = {'key': self.key, 'nodeid': self.node_id})
             if resp.status_code != 200:
                 print('No available tasks so far')
-                return False, None
+                return False
         
             print('Retrieved a task!')
             ###>> retrieving file
@@ -110,7 +129,7 @@ class Node:
                 try:
                     arch = resp.json()['archive_name']
                     fresp = requests.get(front + '/getfile', data = {'key': self.key, 'nodeid': self.node_id, 'archive_name': arch}, stream = True)
-                    assert_response(fresp)
+                    assert_response(fresp, None)
                     print('Retrieving archive!')
                     download_file(fresp, os.path.join(config['ARCHIVE_DIRECTORY'], '.'.join([arch, ARCHIVE_EXTENTION])))
                     break
@@ -122,6 +141,18 @@ class Node:
         else:
             print('Found unfinished task')
             clear_directory(config['WORKING_DIRECTORY'])
+            ###>> extracting max waiting time from lockfile
+            try:
+                ifs = open(tasklockfile, 'r')
+                smaxtime = ifs.read()
+                self.max_time = int(smaxtime)
+            except Exception as e:
+                requests.put(front + '/nodes/{0}'.format(self.node_id), data = {'key': self.key, 'status': 'failed', 'message': str(e)})
+                assert_response(resp, ETaskRecoverable) # It is recoverable unless we can properly reply that it is not
+                raise ETaskUnrecoverable('Malformed task lockfile: can not parse <{0}> as int'.format(smaxtime))
+            finally:
+                ifs.close()
+            ###<< extracting max waiting time from lockfile
 
         ###>> unzipping
         try:
@@ -133,30 +164,27 @@ class Node:
                 raise Exception('Failed to extract: {0}'.format(unzipcmd))
             print('  Done')
         except Exception as e:
-            print('Failed: {0}'.format(str(e)))
             requests.put(front + '/nodes/{0}'.format(self.node_id), data = {'key': self.key, 'status': 'failed', 'message': str(e)})
-            print('  Clearing working and archive directories')
-            clear_directory(config['ARCHIVE_DIRECTORY'])
-            return False, None
+            assert_response(resp, ETaskRecoverable) # It is recoverable unless we can properly reply that it is not
+            raise ETaskUnrecoverable(str(e))
         ###<< unzipping
 
-        try:
+        ###>> creating lockfile
+        if not os.path.exists(tasklockfile):
             print('Creating task lock file: {0}'.format(tasklockfile))
             ofs = open(tasklockfile, 'w')
-            ofs.write('Make it so!')
+            ofs.write(str(self.max_time))
             ofs.close()
-        except Exception as e:
-            print('Failed: {0}'.format(str(e)))
-            print('Proceeding to next step')
+        ###<< creating lockfile
 
         ###>> finding scripts
         start_script = find_script(START_SCRIPT_PREFIX, config['WORKING_DIRECTORY'], config['SCRIPT_VALID_POSTFIX_FMTS'])
         if not (start_script):
-            msg = 'Failed: Not found start script: start = {0}'.format(start_script)
-            print(msg)
+            msg = 'Not found start script: start = {0}'.format(start_script)
             requests.put(front + '/nodes/{0}'.format(self.node_id), data = {'key': self.key, 'status': 'failed', 'message': msg})
-            return False, None
-        ###>> finding scripts
+            assert_response(resp, ETaskRecoverable) # It is recoverable unless we can properly reply that it is not
+            raise ETaskUnrecoverable(msg)
+        ###< finding scripts
 
         ###>> starting
         try:
@@ -165,36 +193,42 @@ class Node:
             self.pdesc = run(start_script)
         except Exception as e:
             msg = 'Failed to run: {0}'.format(str(e))
-            print(msg)
             requests.put(front + '/nodes/{0}'.format(self.node_id), data = {'key': self.key, 'status': 'failed', 'message': msg})
-            return False, None
+            assert_response(resp, ETaskRecoverable) # It is recoverable unless we can properly reply that it is not
+            raise ETaskUnrecoverable(msg)
         ###<< starting
 
         ###>> watchdog
         ### bark-bark!
         print('Setting watchdog on PID = {0}'.format(self.pdesc.pid))
         try:
-            w = Watchdog(n.pdesc)
+            w = Watchdog(self.pdesc, self.max_time)
             w.start()
             w.wait()
             print('Good boy!')
         except Exception as e:
             msg = 'Watchdog is sick: {0}'.format(str(e))
             print(msg)
-            raise Exception(msg)
+            raise EUnrecoverable(msg)
         ###<< wahchdog
 
         if self.pdesc.returncode:
-            print('Task failed with code: {0}, but we will return what we can'.format(self.popen.returncode))
+            msg = 'Task failed with code: {0}'.format(self.pdesc.returncode)
+            print(msg + ', but we will return what we can')
+            requests.put(front + '/nodes/{0}'.format(self.node_id), data = {'key': self.key, 'status': 'failed', 'message': msg})
+            assert_response(resp, ETaskRecoverable)
+            self.status = TASK_STATUS_FAILED
+        else:
+            self.status = TASK_STATUS_FINISHED
 
         ###>> task done, find results
         os.chdir(config['WORKING_DIRECTORY'])
         resdir = os.path.join(config['WORKING_DIRECTORY'], SCRIPT_RESULTS_DIR)
         if not os.path.exists(resdir):
-            msg = 'Failed: results directory not found'
-            print(msg)
+            msg = 'Results directory <{0}> not found in archive root'.format(SCRIPT_RESULTS_DIR)
             requests.put(front + '/nodes/{0}'.format(self.node_id), data = {'key': self.key, 'status': 'failed', 'message': msg})
-            raise Exception(msg)
+            assert_response(resp, ETaskRecoverable) # It is recoverable unless we can properly reply that it is not
+            raise ETaskUnrecoverable(msg)
         
         ### results dir exists, zip it
         os.chdir(resdir)
@@ -202,57 +236,51 @@ class Node:
         newarchfile = os.path.join(config['ARCHIVE_DIRECTORY'], newarchfn)
         zipcmd = ZIPPER_CMD_FMT.format(name = newarchfile, wildcard = '*')
         if os.system(zipcmd):
-            raise Exception('Failed to compress: {0}'.format(zipcmd))
+            raise ETaskUnrecoverable('Failed to compress: {0}'.format(zipcmd))
         os.chdir(config['WORKING_DIRECTORY'])
 
         ### send it
         newarchfile = '.'.join([newarchfile, ARCHIVE_EXTENTION])
         front = 'http://{0}'.format(config['NODE_FRONTEND'])
-        resp = requests.post(front + '/tasks', files = {'file': open(newarchfile, 'rb')}, data = {'key': self.key, 'nodeid': self.node_id})
-        if resp.status_code != 200:
-            raise Exception('Failed: {0}'.format(str(e)))
+        resp = requests.post(front + '/tasks', files = {'file': open(newarchfile, 'rb')}, data = {'key': self.key, 'nodeid': self.node_id, 'status': self.status})
+        assert_response(resp, ETaskRecoverable)
 
         os.unlink(tasklockfile)
         os.chdir(config['WORKING_DIRECTORY'])
-        clear_directory(config['ARCHIVE_DIRECTORY'])
-        clear_directory(config['WORKING_DIRECTORY'])
 
-        print('WOOOOOOOAH! WOOOOOAH! MASAKA! IMPOSSIBURU! STILL NOTHING FAILED, SRSLY?!')
-        print('THEN:')
-        print('DISCONNECT THE CIRCUITS!')
-        print('rm -rf /')
-        print('shutdown --explode --timer=0')
-        print('RIP {0}'.format(datetime.utcnow()))
-
-        return True, None
+        return True
 
 class Watchdog:
     pdesc     = None
     timestamp = None
-    timeout   = 2 # days
+    timeout   = None
     cycletime = 60 # 1 minutes
     running   = False
     kill_to_term = 10 # seconds
 
-    def __init__(self, pdesc):
+    def __init__(self, pdesc, timeout):
         self.pdesc = pdesc
+        self.timeout = timeout
         self.timestamp = datetime.utcnow()
 
     def check(self):
-        if self.pdesc.poll():
-            if (datetime.utcnow() - self.timestamp).days > self.timeout:
+        print('check')
+        if self.pdesc.poll() == None:
+            print('None')
+            if (datetime.utcnow() - self.timestamp).seconds > self.timeout:
                 self.pdesc.terminate()
                 sleep(self.kill_to_term)
                 if self.pdesc.poll():
                     self.pdesc.kill()
             else:
+                print('Bark!')
                 self.running = True
                 self.start()
+        print('not None')
         self.running = False
-        print('Bark!')
-
 
     def start(self):
+        print('Start')
         self.running = True
         thr = threading.Timer(self.cycletime, self.check)
         thr.daemon = True
@@ -261,6 +289,14 @@ class Watchdog:
     def wait(self):
         while self.running:
             sleep(self.cycletime)
+
+def assert_response(resp, ExcClass, code = 200):
+    if resp.status_code != code:
+        msg = 'Response was not {0}: {1}'.format(code, resp.text)
+        if ExcClass:
+            raise ExcClass(msg)
+        else:
+            print(msg)
 
 ###
 def jsr(method, address, data = {}):
@@ -284,10 +320,6 @@ def run(cmd):
     else:
         raise Exception('Executing on OS {0} is currently not supported'.format(sysname))
 
-def assert_response(resp, code = 200):
-    if resp.status_code != code:
-        raise Exception('Response was not {0}: {1}'.format(code, resp.json()))
-
 def random_string(size):
     return ''.join([random.choice(string.ascii_letters) for i in range(size)])
 
@@ -302,7 +334,7 @@ def download_file(resp, fullpath):
 
 def find_script(prefix, path, valid_fmts):
     print('  Searching for start script:')
-    for fmt in valid_fmts:
+    for fmt in valid_fmts + [START_SCRIPT_PREFIX]:
         name = PLATFORM_INFO_SEP.join([\
             prefix,\
             fmt.format(\
@@ -325,25 +357,40 @@ def create_dir_if_not_exists(path):
         print('Creating directory: {0}'.format(path))
         os.makedirs(path)
  
-def clear_directory(folder, whitelist=[config['ARCHIVE_DIRECTORY']]):
+def clear_directory(folder):
     for the_file in os.listdir(folder):
         file_path = os.path.join(folder, the_file)
         try:
             if os.path.isfile(file_path):
                 os.unlink(file_path)
             elif os.path.isdir(file_path):
-                if file_path not in map(lambda fld: os.path.split(fld)[-1], whitelist):
-                    shutil.rmtree(file_path)
+                shutil.rmtree(file_path)
         except Exception as e:
             print(str(e))
 
 def parse_cli_argv():
     try:
-        node_front = sys.argv[1]
+        cli_num, node_front = sys.argv[1], sys.argv[2]
+        if len(sys.argv) == 4:
+            traits_file = sys.argv[3]
+            try:
+                print('Parsing traits:')
+                with open(traits_file, 'r') as ifs:
+                    for line in ifs:
+                        ln = ''.join(filter(lambda s: s not in ['\r', '\n'], line))
+                        if len(ln) > 0:
+                            name, ver = ln.split(TRAITS_COLSEP, 1)
+                            config['CLIENT_TRAITS'] += [dict(name = name, version = ver)]
+                            print('  name: "{0}", version = "{1}"'.format(name, ver))
+            except Exception as e:
+                print('Traits syntax error {0}'.format(str(e)))
+                raise e
     except Exception as e:
-        print('Usage: {0} node_frontend_host:node_frontend_port')
+        print('Usage: {0} number node_frontend_host:node_frontend_port [traits_file]')
+        print('Traits name and version must be separated with single space character.')
+        print('Everything after the first space is treated as version.')
         sys.exit()
-    return node_front
+    return cli_num, node_front
 
 def set_lowest_priority():
     sysname = platform.system()
@@ -356,7 +403,13 @@ def set_lowest_priority():
 if __name__ == '__main__':
     print(config['GRID_CALC_ROLE'])
 
-    config['NODE_FRONTEND'] = parse_cli_argv()
+    node_num, config['NODE_FRONTEND'] = parse_cli_argv()
+    config['GRID_CALC_ROLE'] = '_'.join([config['GRID_CALC_ROLE'], str(node_num)])
+    config.update(WORKING_DIRECTORY = os.path.join(tempfile.gettempdir(), '_'.join([config['GRID_CALC_ROLE'], 'work'])))
+    config.update(HOME_DIRECTORY = os.path.join(os.path.expanduser("~"), config['GRID_CALC_ROLE']))
+    config.update(ARCHIVE_DIRECTORY = os.path.join(tempfile.gettempdir(), '_'.join([config['GRID_CALC_ROLE'], 'arch'])))
+    config.update(TASK_LOCKFILE = os.path.join(config['HOME_DIRECTORY'], 'task.lock'))
+    config.update(KEYFILE = os.path.join(config['HOME_DIRECTORY'], 'key.dat'))
 
     create_dir_if_not_exists(config['WORKING_DIRECTORY'])
     create_dir_if_not_exists(config['ARCHIVE_DIRECTORY'])
@@ -366,8 +419,26 @@ if __name__ == '__main__':
     set_lowest_priority()
     
     n = Node()
-    try:
-        while True:
-            n.run()
-    except Exception as e:
-        print('Error: {0}'.format(str(e)))
+    while True:
+        try:
+            if n.run():
+                clear_directory(config['WORKING_DIRECTORY'])
+                clear_directory(config['ARCHIVE_DIRECTORY'])
+            else:
+                sleep(n.task_req_timeout)
+        except ETaskRecoverable as tre:
+            print('Retring task: {0}, retrying'.format(str(tre)))
+            clear_directory(config['WORKING_DIRECTORY'])
+        except ETaskUnrecoverable as ture:
+            print('Task is dropped: {0}'.format(str(ture)))
+            clear_directory(config['WORKING_DIRECTORY'])
+            clear_directory(config['ARCHIVE_DIRECTORY'])
+            if os.path.exists(config['TASK_LOCKFILE']):
+                os.unlink(config['TASK_LOCKFILE'])
+        except ERecoverable as re:
+            print('Restarting: {0}'.format(str(re)))
+        except (EUnrecoverable, Exception) as ue:
+            print('Unrecoverable error: {0}'.format(str(ue)))
+            clear_directory(config['WORKING_DIRECTORY'])
+            clear_directory(config['ARCHIVE_DIRECTORY'])
+            break
